feat: add new components and adjust existing structures
