+++ src/components/PastAnal/CandidateSummary.vue
@@ -1,4 +1,7 @@
 <template>
+  <h4 class="h4 mb-8">
+    <i class="bi bi-person-raised-hand me-2"></i>候選人情況
+  </h4>
   <div class="mb-8">
     <div class="d-flex justify-content-center py-6 mb-8a">
       <PastAnalPie id="past-candidate-summary" :data="pieData"></PastAnalPie>
